Move metabox within the Embeddings class
