+++ includes/Classifai/Plugin.php
@@ -135,7 +135,6 @@ class Plugin {
 			'debug_info'    => new Admin\DebugInfo(),
 			'bulk_actions'  => new Admin\BulkActions(),
 			'updater'       => new Admin\Update(),
-			'metabox'       => new Admin\Metabox(),
 		];
 
 		foreach ( $this->admin_helpers as $instance ) {
